Fix artifacting of infobar
This was a doozy. The problem was that drawStringWithShadow would use a different texture from the normal one, and minecraft would not replace the texture for the rest of it's UI elements, therefore causing the textures to be replaced to letters. Additionally, forcing it to be called only when the event is building text can improve performance by reducing the amount of loops it needs to perform, while also holding back error cases like above.